Update Header.tsx
The Link component below the div below Header is used to display the logo
The Link makes it clickable to return back to the webpage.

The div below the Link has a justify end to position the login button at the end of the navbar.

This div contains our login buttons and profile button for logging out