Created user.actions.ts
The createUserParams import was created to specify the parameters to be passed into the object.
The handleError was a utility function created to handle errors
The connectToDatabase function was a function we created to setup a connection with the MongoDB database
The User Model is a schema model for the User Object

In the createUser Function, we are passing an object called user that fetches its params from CreateUserParams. The function then fires connectToDatabase function.
if it is successful, it creates a new User using the User Schema Model and parses the data using JSON

If there is any error, it is caught using our utility function