Create page.tsx
This is the page that will be navigated to when the user goes to the url localhost:3000/events/create.  Apart from the semantic section and h3 tags, the main component to be considered here is the EventForm. The EventForm is a shared component that will contain all of our fields