Created Dropdown.tsx
The dropdown is a custom component made by combining two components from shadcn i.e The Select Component and the Alert Dialog Component. We have opted for this approach because we want to allow the user to choose from default categories and also have the ability to make his own category.

In the Select Component, In the component called SelectContent, we are using react useState to map over all the categories in our backend. Since we don't have a backend yet, currently the component doesn't map over anything as such we have initialized it to an empty array.
We are using the react state  variable categories and  function setCategories for this feature.

After the list of categories, we have used the alert dialog component from shadcn. This component will create an alert dialog box where the user can specify a title for a new category. The react state variable newCategory and function setNewCategory will then be used to add this category to our backend