Update lib\actions\event.actions.ts
Created a new event action called getAllEvents. This action will be used to fetch all the events from the database. This action will also later be used to implement filtering and to find related events using conditions. The conditions have not been defined yet.
For now we are fetching all the events.

The eventsQuery has parameters for sort to get the most recent events, to not skip any events and to only show the number of events specified in the limit.

The eventscount variable uses the countDocuments method to count the number of documents that will be fetched. This will be used to implement pagination.

In this method, we are returning an object, instead of returning just the events