Update Collection.tsx
In the collection component we are first checking if there are more than 0 events in our database.
If there are, we map over each event, for every event, we  check if the user has already purchased the event, For that we have the 2 variables hasOrderLink and hidePrice

and then we return a Card Component for each event. The Card is a custom component developed to display the details of each event.

IF there are no events, we just return emptyTitle and emptyStateSubtext