Create components\shared\Card.tsx
In the Card component, we are first fetching the userId using Clerk. We will then use that userId to check if that user is creator of that event.
If he is, then we display an edit button. The edit button will redirect him to the update page where he can change the details of the Event.
After that validation, we have divs and spans to display the details of the event. These details will be available to everyone even those who aren't logged in