Update app\(root)\events\[id]\update\page.tsx
Defined the UpdateEventProps which is just an id with a string value.
Then we are fetching the event using the GetEventByID method. The event is then passed to the EventForm as props.
We are also using clerk's sessionClaims to get the userId.
This userId is used to verify if this user is authorized to make the update in the Eventform