Update app\(root)\events\[id]\page.tsx
We are now also fetching related Events on our main Event page. We are doing this by first fetching the related Events using the event Action. This event data is then passed into the collection component. The collection is a reusable component so we don't have to do anything else